Complete refactor of FishLeg repo. Got rid of lots of extra methods. Aux update not working yet.
