+++ src/optim/FishLeg/fishleg.py
@@ -2,26 +2,9 @@ from typing import Tuple, Callable, Any, Union, List
 from collections.abc import Mapping
 import torch
 import torch.nn as nn
-import numpy as np
-from torch.nn import init
 from torch.optim import Optimizer, Adam
-import sys
-import regex as re
-import csv
-import os
-from transformers.models.bert.modeling_bert import BertAttention
-
-from .utils import (
-    recursive_setattr,
-    recursive_getattr,
-    update_dict,
-    get_named_layers_by_regex,
-    NamedLayer,
-)
-from transformers import get_scheduler
 
 from .layers import *
-from .fishleg_layers import *
 from .fishleg_likelihood import FishLikelihood
 
 __all__ = [
@@ -78,8 +61,6 @@ class FishLeg(Optimizer):
         >>>
         >>> likelihood = FixedGaussianLikelihood(sigma=1.0)
         >>>
-        >>>
-        >>>
         >>> model = nn.Sequential(
         >>>     nn.Linear(2, 5),
         >>>     nn.ReLU(),
@@ -119,7 +100,6 @@ class FishLeg(Optimizer):
         update_aux_every: int = 10,
         warmup_steps: int = 0,
         normalize_aux: bool = False,
-        module_names: str = "__ALL__",  # This should be __ALL__ or something similar
         initialization: str = "uniform",
         grad_clip: bool = False,
         device: str = "cpu",
@@ -130,14 +110,6 @@ class FishLeg(Optimizer):
         self.aux_dataloader = aux_dataloader
         self.likelihood = likelihood
 
-        self.lr = lr
-        self.beta = beta
-
-        self.weight_decay = weight_decay
-
-        self.aux_lr = aux_lr
-        self.fish_scale = fish_scale
-        self.damping = damping
         self.update_aux_every = update_aux_every
 
         self.warmup_steps = warmup_steps
@@ -149,21 +121,31 @@ class FishLeg(Optimizer):
         self.device = device
         self.verbose = verbose
 
-        self.model, param_groups = self.init_model_aux(
-            model, module_names=module_names
+        # TODO: Add more in here?
+        defaults = dict(
+            lr=lr,
+            beta=beta,
+            weight_decay=weight_decay,
+            aux_lr=aux_lr,
+            fish_scale=fish_scale,
+            damping=damping,
         )
-        defaults = dict(aux_lr=aux_lr, lr=lr)
-        super(FishLeg, self).__init__(param_groups, defaults)
+        params = [
+            param
+            for name, param in model.named_parameters()
+            if "fishleg_aux" not in name
+        ]
+        super(FishLeg, self).__init__(params, defaults)
 
-        aux_param = [
+        aux_params = [
             param for name, param in model.named_parameters() if "fishleg_aux" in name
         ]
-        if self.likelihood is not None:
-            if len(self.likelihood.get_parameters()) > 0:
-                aux_param.extend(self.likelihood.get_aux_parameters())
+
+        if len(self.likelihood.get_parameters()) > 0:
+            aux_params.extend(self.likelihood.get_aux_parameters())
 
         self.aux_opt = Adam(
-            aux_param,
+            aux_params,
             lr=aux_lr,
             betas=aux_betas,
             eps=aux_eps,
@@ -171,206 +153,40 @@ class FishLeg(Optimizer):
         )
 
         if self.warmup_steps > 0:
-            self.warmup_aux(fish_scale=fish_scale)
+            self.warmup_aux()
 
         self.step_t = 0
         self.store_g = True
 
-    # I think we should demand a Fisher model be constructed outside the optimizer before being passed in.
-    def init_model_aux(
-        self,
-        model: nn.Module,
-        module_names: str,
-    ) -> Union[nn.Module, List]:
-        """Given a model to optimize, parameters can be devided to
-
-        #. those fixed as pre-trained.
-        #. those required to optimize using FishLeg.
-
-        Replace modules in the second group with FishLeg modules.
-
-        Args:
-            model (:class:`torch.nn.Module`, required):
-                A model containing modules to replace with FishLeg modules
-                containing extra functionality related to FishLeg algorithm.
-        Returns:
-            :class:`torch.nn.Module`, the replaced model.
-        """
-        if any(["weight" in m for m in module_names]):
-            raise TypeError(
-                f"Parameters to be optimized in FishLeg are considered together in one module, and cannot be optimized individually"
-            )
-
-        # Add auxiliary parameters
-        if module_names == "__ALL__":
-            named_layers = [
-                NamedLayer(name, layer) for name, layer in model.named_modules()
-            ]
-
-        else:
-            named_layers = get_named_layers_by_regex(model, module_names)
-
-        replaced_layers = []
-        for named_layer in named_layers:
-            module = named_layer.layer
-            module_name = named_layer.layer_name
-
-            for layer in replaced_layers:
-                if re.match(layer.layer_name, module_name):
-                    inner_name = module_name[len(layer.layer_name) + 1 :]
-                    if any(
-                        re.match(inner_name, param_name)
-                        for param_name in layer.layer.order
-                    ):
-                        continue
-
-            if isinstance(module, nn.Linear):
-                replace = FishLinear(
-                    module.in_features,
-                    module.out_features,
-                    module.bias is not None,
-                    device=self.device,
-                )
-
-                # if self.initialization == "normal":
-                #     init.normal_(replace.weight, 0, 1 / np.sqrt(module.in_features))
-                # elif self.initialization == "zero":  # fill with zeros for adapters
-                #     module.weight.data.zero_()
-                #     module.bias.data.zero_()
-            
-            elif isinstance(module, nn.Embedding):
-                replace = FishEmbedding(
-                    num_embeddings=module.num_embeddings,
-                    embedding_dim=module.embedding_dim,
-                    padding_idx=module.padding_idx,
-                    max_norm=module.max_norm,
-                    norm_type=module.norm_type,
-                    scale_grad_by_freq=module.scale_grad_by_freq,
-                    sparse=module.sparse,
-                    device=self.device,
-                )
-            
-            elif isinstance(module, nn.Conv2d):
-                replace = FishConv2d(
-                    in_channels=module.in_channels,
-                    out_channels=module.out_channels,
-                    kernel_size=module.kernel_size,
-                    stride=module.stride,
-                    padding=module.padding,
-                    dilation=module.dilation,
-                    groups=module.groups,
-                    bias=(module.bias is not None),
-                    padding_mode=module.padding_mode,
-                    device=self.device
-                    # TODO: deal with dtype and device?
-                )
-            
-            elif isinstance(module, BertAttention):
-                config = model.config
-                replace = FishBertAttention(config, device=self.device)
-            
-            elif isinstance(module, nn.BatchNorm2d):
-                replace = FishBatchNorm2d(
-                    num_features=module.num_features,
-                    eps=module.eps,
-                    momentum=module.momentum,
-                    affine=module.affine,
-                    track_running_stats=module.track_running_stats,
-                    init_scale=self.fish_scale,
-                    device=self.device,
-                )
-            
-            elif isinstance(module, nn.LayerNorm):
-                replace = FishLayerNorm(
-                    normalized_shape=module.normalized_shape,
-                    eps=module.eps,
-                    elementwise_affine=module.elementwise_affine,
-                    init_scale=self.fish_scale,
-                    device=self.device,
-                )
-            else:
-                continue
-                
-            replace = update_dict(replace, module)
-            recursive_setattr(model, module_name, replace)
-            replaced_layers.append(NamedLayer(module_name, replace))
-            if self.verbose:
-                print("Replaced with FishLeg Layer: \t ", module_name)
-
-        # Define each modules
-        param_groups = []
-        for named_layer in replaced_layers:
-            module = named_layer.layer
-            params = {
-                name: param
-                for name, param in module.named_parameters()
-                if "fishleg_aux" not in name
-            }
-
-            g = {
-                "params": [params[name] for name in module.order],
-                "gradbar": [torch.zeros_like(params[name]) for name in module.order],
-                "grad": [torch.zeros_like(params[name]) for name in module.order],
-                "u": [torch.zeros_like(params[name]) for name in module.order],
-                "Qv": module.Qv,
-                "order": module.order,
-                "name": named_layer.layer_name,
-                "module": module,
-            }
-            param_groups.append(g)
-
-        if self.likelihood is not None:
-            likelihood_params = self.likelihood.get_parameters()
-            if len(likelihood_params) > 0:
-                self.likelihood.init_aux(init_scale=self.fish_scale)
-                g = {
-                    "params": likelihood_params,
-                    "gradbar": [torch.zeros_like(p) for p in likelihood_params],
-                    "theta0": [p.clone() for p in likelihood_params],
-                    "grad": [torch.zeros_like(p) for p in likelihood_params],
-                    "Qv": self.likelihood.Qv,
-                    "order": self.likelihood.order,
-                    "name": "likelihood",
-                }
-                param_groups.append(g)
-
-        # TODO: The above may not be a very "correct" way to do this, so please feel free to change, for example, we may want to check the name is in the fish_layer keys before attempting what is in the try statement.
-        # TODO: Error checking to check that model includes some auxiliary arguments.
-
-        return model, param_groups
-
-    def warmup_aux(
-        self,
-        dataloader: torch.utils.data.DataLoader,
-        loss: Callable[
-            [nn.Module, Tuple[torch.Tensor, torch.Tensor]], torch.Tensor
-        ] = None,
-    ) -> None:
+    def warmup_aux(self) -> None:
         """
         Warm up auxilirary parameters with approxiamte Adam
         """
-        for _ in range(0, self.warmup_steps, dataloader.batch_size):
-            data = self._prepare_input(next(iter(dataloader)))
-            loss(self.model, data).backward()
-            self._store_u(transform=lambda x: x * x)
+        # TODO: Add checking for this function!
+        # TODO: The below loop can be refactored.
+        for _ in range(0, self.warmup_steps, self.aux_dataloader.batch_size):
+            data = self._prepare_input(next(iter(self.aux_dataloader)))
+            output = self.model(data[0])
+            self.likelihood(output, data[1]).backward()
 
-        for group in self.param_groups:
-            ds = []  # What is this?
-            for i, g2 in enumerate(group["grad"]):
-                g2_avg = g2 / int(self.warmup_steps / dataloader.batch_size)
-                ds.append(
-                    self.fish_scale / torch.sqrt(torch.sqrt(g2_avg + self.damping))
-                )
+        group = self.param_groups[0]
+        fish_scale = group["fish_scale"]
+        damping = group["damping"]
 
-            group["module"].warmup(ds, init_scale=self.fish_scale)
+        for module in self.model.modules():
+            warm_facs = []
+            if not isinstance(module, nn.Sequential) and hasattr(module, "warmup"):
+                for name, param in module.named_parameters():
+                    if "fishleg_aux" not in name:
+                        g_avg = param.grad / int(
+                            self.warmup_steps / self.aux_dataloader.batch_size
+                        )
+                        warm_facs.append(fish_scale / (torch.square(g_avg) + damping))
+                module.warmup(warm_facs, init_scale=fish_scale)
+            else:
+                continue
 
-    def parameters(
-        self,
-    ):
-        for group in self.param_groups:
-            for param in group["params"]:
-                yield param
-
+    # TODO: ALL of this.
     def pretrain_fish(
         self,
         dataloader: torch.utils.data.DataLoader,
@@ -408,39 +224,7 @@ class FishLeg(Optimizer):
 
         return info
 
-    # This should be external
-    def _store_u(
-        self,
-        transform: Callable = lambda x: x,
-        alpha: float = 1.0,
-        new: bool = False,
-    ):
-        for group in self.param_groups:
-            for i, p in enumerate(group["params"]):
-                grad = transform(p.grad.data)
-
-                if not new:
-                    group["grad"][i].add_(grad, alpha=alpha)
-                else:
-                    group["grad"][i].copy_(grad)
-                group["u"][i].copy_(grad)
-
-    # This should be external and done before passing into the optimizer
-    def _prepare_input(
-        self, data: Union[torch.Tensor, Any]
-    ) -> Union[torch.Tensor, Any]:
-        """
-        Prepares one `data` before feeding it to the model, be it a tensor or a nested list/dictionary of tensors.
-        """
-        if isinstance(data, Mapping):
-            return type(data)({k: self._prepare_input(v) for k, v in data.items()})
-        elif isinstance(data, (tuple, list)):
-            return type(data)(self._prepare_input(v) for v in data)
-        elif isinstance(data, torch.Tensor):
-            kwargs = dict(device=self.device)
-            return data.to(**kwargs)
-        return data
-
+    # TODO: ALL of this.
     def update_aux(self) -> None:
         """
         Performs a single auxliarary parameter update
@@ -456,6 +240,9 @@ class FishLeg(Optimizer):
         data = next(iter(self.aux_dataloader))
         data = self._prepare_input(data)
 
+        group = self.param_groups[0]
+        damping = group["damping"]
+
         self.aux_opt.zero_grad()
 
         data_x, data_y = data
@@ -470,7 +257,7 @@ class FishLeg(Optimizer):
 
         self.zero_grad()
         self.likelihood.nll(pred_y, samples_y).backward()
-        
+
         if self.grad_clip:
             nn.utils.clip_grad_norm_(
                 self.parameters(),
@@ -487,10 +274,10 @@ class FishLeg(Optimizer):
 
             for p, g, d_p in zip(group["params"], group["grad"], qg):
                 grad = p.grad.data
-            
+
                 quad_term = quad_term + torch.sum(grad * d_p)
                 linear_term = linear_term + torch.sum(g * d_p)
-                reg_term = reg_term + self.damping * torch.sum(d_p * d_p)
+                reg_term = reg_term + damping * torch.sum(d_p * d_p)
                 align = align + torch.sum(grad * g)
 
         check = quad_term * align + self.damping * linear_term - g2
@@ -516,77 +303,128 @@ class FishLeg(Optimizer):
         self.store_g = True
         return aux_loss, check, linear_term, quad_term, reg_term, g2
 
-    def step(self, closure=None) -> None:
+    # TODO: What do we need here?
+    def _init_group(
+        self,
+        group,
+        # params_with_grad,
+        # grads,
+        exp_avgs,
+        # exp_avg_sqs,
+        # max_exp_avg_sqs,
+        state_steps,
+    ):
+        for p in group["params"]:
+            if p.grad is not None:
+                # params_with_grad.append(p)
+                # if p.grad.is_sparse:
+                #     raise RuntimeError(
+                #         "Adam does not support sparse gradients, please consider SparseAdam instead"
+                #     )
+                # grads.append(p.grad)
+
+                state = self.state[p]
+                # Lazy state initialization
+                if len(state) == 0:
+                    state["step"] = (
+                        torch.zeros((1,), dtype=torch.float, device=p.device)
+                        # if group["capturable"] or group["fused"]
+                        # else torch.tensor(0.0)
+                    )
+                    # Exponential moving average of gradient values
+                    state["exp_avg"] = torch.zeros_like(
+                        p, memory_format=torch.preserve_format
+                    )
+                    # Exponential moving average of squared gradient values
+                    # state["exp_avg_sq"] = torch.zeros_like(
+                    #     p, memory_format=torch.preserve_format
+                    # )
+                    # if group["amsgrad"]:
+                    #     # Maintains max of all exp. moving avg. of sq. grad. values
+                    #     state["max_exp_avg_sq"] = torch.zeros_like(
+                    #         p, memory_format=torch.preserve_format
+                    #     )
+
+                exp_avgs.append(state["exp_avg"])
+                # exp_avg_sqs.append(state["exp_avg_sq"])
+
+                # if group["amsgrad"]:
+                #     max_exp_avg_sqs.append(state["max_exp_avg_sq"])
+                # if group["differentiable"] and state["step"].requires_grad:
+                #     raise RuntimeError(
+                #         "`requires_grad` is not supported for `step` in differentiable mode"
+                #     )
+                state_steps.append(state["step"])
+
+    # TODO: Add in the aux update.
+    def step(self) -> None:
         """
         Performes a single optimization step of FishLeg.
         """
         self.updated = False
 
-        if self.update_aux_every > 0:
-            if self.step_t % self.update_aux_every == 0:
-                self._store_u(new=True)
+        # if self.step_t % self.update_aux_every == 0:
+        #     self._store_u(new=True)
 
-            if self.step_t % self.update_aux_every == 1:
-                # once for difference of gradients
-                self._store_u(alpha=-1.0)
-                info = self.update_aux()
-                # info = [e.detach().cpu().numpy() for e in info]
+        # if self.step_t % self.update_aux_every == 1:
+        #     # once for difference of gradients
+        #     self._store_u(alpha=-1.0)
+        #     info = self.update_aux()
 
-                # if self.verbose == True:
-                #     if self.step_t % 200 == 1:
-                #         print(
-                #             "iter:{:d}, lr:{:.2f} \tauxloss:{:.2f} \tcheck:{:.2f} \tlinear:{:.2f} \tquad:{:.2f} \treg:{:.2f} \tg2:{:.2f}".format(
-                #                 self.step_t, self.fish_lr, *info
-                #             )
-                #         )
+        #     self.updated = True
 
-                self.updated = True
-
-            if self.step_t % self.update_aux_every == 2:
-                # once for gradient
-                self._store_u(new=True)
-                self.update_aux()
-                self.updated = True
-            
-        elif self.update_aux_every < 0:
-            self._store_u(new=True)
-
-            for _ in range(-self.update_aux_every):
-                self.update_aux()
-
-            self.updated = True
+        # if self.step_t % self.update_aux_every == 2:
+        #     # once for gradient
+        #     self._store_u(new=True)
+        #     self.update_aux()
+        #     self.updated = True
 
         self.step_t += 1
 
-        for group in self.param_groups:
-            name = group["name"]
+        group = self.param_groups[0]
 
-            with torch.no_grad():
-                nat_grad = group["Qv"](
-                    group["u"]
-                    if self.updated
-                    else [p.grad.data for p in group["params"]]
-                )
-            
-                for p, d_p, gbar in zip(group["params"], nat_grad, group["gradbar"]):
-                    gbar.copy_(self.beta * gbar + (1.0 - self.beta) * d_p)
-                    delta = gbar.add(p, alpha=self.weight_decay)
-                    p.add_(delta, alpha=-self.lr)
+        # TODO: get rid of anything we don't need
+        # params_with_grad = []
+        # grads = []
+        exp_avgs = []
+        # exp_avg_sqs = []
+        # max_exp_avg_sqs = []
+        state_steps = []
+        beta = group["beta"]
+        weight_decay = group["weight_decay"]
+        step_size = group["lr"]  # TODO: Add scheduling in here.
 
-    @torch.no_grad()
-    def _save_input(
-        self,
-        module: torch.nn.Module,
-        input_: List[torch.Tensor],
-    ) -> None:
-        if self.store_g and module.training:
-            module.save_layer_input(input_)
+        self._init_group(
+            group,
+            # params_with_grad,
+            # grads,
+            exp_avgs,
+            # exp_avg_sqs,
+            # max_exp_avg_sqs,
+            state_steps,
+        )
 
-    @torch.no_grad()
-    def _save_grad_output(
-        self,
-        module: torch.nn.Module,
-        grad_output: Union[Tuple[torch.Tensor, ...], torch.Tensor],
-    ) -> None:
-        if self.store_g and module.training:
-            module.save_layer_grad_output(grad_output)
+        # TODO: Could do with a better way of indexing the exp_avgs
+        with torch.no_grad():
+            p_idx = 0
+            for module in self.model.modules():
+                if not isinstance(module, nn.Sequential) and hasattr(module, "Qv"):
+                    layer_grads = []
+                    for name, param in module.named_parameters():
+                        if "fishleg_aux" not in name:
+                            layer_grads.append(param.grad.data)
+
+                    nat_grads = module.Qv(layer_grads)
+
+                    for n, (name, param) in enumerate(module.named_parameters()):
+                        if "fishleg_aux" not in name:
+                            nat_grad = nat_grads[n]
+                            exp_avg = exp_avgs[p_idx]
+                            p_idx += 1
+
+                            exp_avg.mul_(beta).add_(nat_grad, alpha=1 - beta)
+
+                            if weight_decay != 0:
+                                exp_avg = exp_avg.add(param, alpha=weight_decay)
+
+                            param.add_(exp_avg, alpha=-step_size)
