+++ src/optim/FishLeg/layers/__init__.py
@@ -1,1 +1,8 @@
-from .bert import *
+from .fish_linear import FishLinear
+from .fish_conv2d import FishConv2d
+from .fish_embedding import FishEmbedding
+from .fish_bert import FishBertAttention
+from .fish_batchNorm2d import FishBatchNorm2d
+from .fish_LayerNorm import FishLayerNorm
+
+from .utils import *
