+++ src/optim/FishLeg/layers/fish_base.py
@@ -1,0 +1,74 @@
+import torch.nn as nn
+from torch import Tensor
+from torch.nn import ParameterDict
+
+from typing import Any, List, Dict, Tuple
+
+from abc import abstractmethod
+
+
+class FishModule(nn.Module):
+    """Base class for all neural network modules in FishLeg to
+
+    #. Initialize auxiliary parameters, :math:`\lambda` and its forms, :math:`Q(\lambda)`.
+    #. Specify quick calculation of :math:`Q(\lambda)v` products.
+
+    :param torch.nn.ParameterDict fishleg_aux: auxiliary parameters
+                with their initialization, including an additional parameter, scale,
+                :math:`\eta`. Make sure that
+
+                .. math::
+                        - \eta_{init} Q(\lambda) grad = - \eta_{sgd} grad
+
+                is hold in the beginning of the optimization
+    :param List order: specify a name order of original parameter
+
+    """
+
+    def __init__(self, *args: Any, **kwargs: Any) -> None:
+        super(FishModule, self).__init__(*args, **kwargs)
+        self.__setattr__("fishleg_aux", ParameterDict())
+        self.__setattr__("order", List)
+
+    @property
+    def name(self) -> str:
+        return self._layer_name
+
+    def cuda(self, device: str) -> None:
+        for p in self.fishleg_aux.values():
+            p.to(device)
+
+    def warmup(
+        self,
+        v: Tuple[Tensor, Tensor] = None,
+        batch_speedup: bool = False,
+        init_scale: float = 1.0,
+    ):
+        pass
+
+    @abstractmethod
+    def Qv(
+        self, aux: Dict, v: Tuple[Tensor, ...], full: bool = False
+    ) -> Tuple[Tensor, ...]:
+        """:math:`Q(\lambda)` is a positive definite matrix which will effectively
+        estimate the inverse damped Fisher Information Matrix. Appropriate choices
+        for :math:`Q` should take into account the architecture of the model/module.
+        It is usually parameterized as a positive definite Kronecker-factored
+        block-diagonal matrix, with block sizes reflecting the layer structure of
+        the neural networks.
+
+        Args:
+            aux: (Dict, required): auxiliary parameters,
+                    :math:`\lambda`, a dictionary with keys, the name
+                    of the auxiliary parameters, and values, the auxiliary parameters
+                    of the module. These auxiliaray parameters will form :math:`Q(\lambda)`.
+            v: (Tuple[Tensor, ...], required): Values of the original parameters,
+                    in an order that align with `self.order`, to multiply with
+                    :math:`Q(\lambda)`.
+            full: (bool, optional), whether to use full inner and outer re-scaling
+        Returns:
+            Tuple[Tensor, ...]: The calculated :math:`Q(\lambda)v` products,
+                    in same order with `self.order`.
+
+        """
+        raise NotImplementedError(f'Module is missing the required "Qv" function')
